Move project banners into own component
